refactor:code and style refactor
