--- packages/react-schema-editor/src/components/FieldEditor.css
@@ -1,16 +1,0 @@
-.field-editor .field-group-title{
-  margin-top: 15px;
-  font-size: 16px;
-}
-.field-editor .field-group-content{
-  display: flex;
-  align-items: flex-start;
-}
-.field-editor .field-group-content .field-group-form-item{
-  width: 20%;
-  margin-right: 0;
-  padding-right: 10px;
-}
-.field-editor .field-group-form-item.auto-complete .ant-form-item-control{
-  margin-top: -1px;
-}
