+++ packages/react-schema-editor/src/index.tsx
@@ -1,22 +1,38 @@
-import React, { useState } from 'react'
-import { Button, Row, Col, Radio } from 'antd'
+import React, {useState} from 'react'
+import { Button, Radio, Tabs } from 'antd'
+import * as fp from 'lodash/fp'
+import _ from 'lodash'
 import { SchemaTree } from './components/SchemaTree'
+import FieldEditor from './components/FieldEditor'
 import { SchemaCode } from './components/SchemaCode'
 import nextComponents from './utils/nextCompProps'
 import antdComponents from './utils/antdCompProps'
-
 import 'antd/dist/antd.css'
+import './main.scss'
 
 export const SchemaEditor: React.FC<{
   schema: any
   onChange: (schema: any) => void
 }> = ({ schema, onChange }) => {
   const [componentType, setComponentType] = useState('antd');
+  const [selectedPath, setSelectedPath] = React.useState(null);
 
-  function handleTypeChange(e) {
+  const handleTypeChange = (e) => {
     setComponentType(e.target.value);
   }
 
+  const handleTreeSelect = (path) => {
+    setSelectedPath(path)
+  }
+
+  const handleCodeChange = (code) => {
+    console.log(code);
+  }
+
+  const selectedSchema =
+  selectedPath &&
+  (selectedPath === 'root' ? schema : fp.get(selectedPath, schema))
+  console.log('selectedSchema', selectedSchema)
   return (
     <div className="schema-editor">
       <div className="schema-menus">
@@ -31,19 +47,35 @@ export const SchemaEditor: React.FC<{
             <Radio value="fusion">Fusion Design组件</Radio>
           </Radio.Group>
         </span>
-
       </div>
-      <Row className="schema-editor-main" style={{ minHeight: '350px' }}>
-        <Col span={24} className="schema-col schema-tree splitter">
-          <SchemaTree schema={schema} onChange={onChange} components={componentType === 'fusion' ? nextComponents : antdComponents}/>
-        </Col>
-        <Col
-          className="schema-col schema-code"
-          style={{ position: 'absolute', right: '0px', height: '100%' }}
-        >
-          <SchemaCode schema={schema} onChange={onChange}></SchemaCode>
-        </Col>
-      </Row>
+      <div className="schema-editor-main">
+        <div className="schema-tree">
+          <SchemaTree schema={schema} onChange={onChange} onSelect ={handleTreeSelect} />
+        </div>
+        <div className="schema-tabs">
+          <Tabs type="card">
+            <Tabs.TabPane tab="属性编辑" key="1">
+              {selectedSchema && <FieldEditor
+                  components={componentType === 'fusion' ? nextComponents : antdComponents}
+                  fieldKey="fieldC"
+                  onFieldKeyChange={value => {
+                    console.log('onFieldKeyChange====', value)
+                  }}
+                  schema={selectedSchema}
+                  onChange={value => {
+                    const newSchema = _.clone(schema)
+                    _.set(newSchema, selectedPath, value)
+                    onChange(newSchema)
+                  }}
+              />}
+            </Tabs.TabPane>
+            <Tabs.TabPane tab="Schema源码" key="2">
+              <SchemaCode schema={schema} onChange={handleCodeChange}></SchemaCode>
+            </Tabs.TabPane>
+            <Tabs.TabPane tab="预览" key="3"></Tabs.TabPane>
+          </Tabs>
+        </div>
+      </div>
     </div>
   )
 }
