--- packages/react-schema-editor/src/utils/index.ts
@@ -1,4 +1,0 @@
-export function jsonToSchema(json?: object) {
-  let schema = json || {}
-  return schema
-}
