+++ packages/react-schema-editor/tsconfig.json
@@ -1,5 +1,6 @@
 {
   "extends": "../../tsconfig.json",
+  "noUnusedLocals": false,
   "compilerOptions": {
     "outDir": "./lib"
   },
